Centralize all assertion handling into AssertionErrorBuilder.
* This avoids the myriad of functions we had scattered about that created error messages used by assertion errors.
* Propagates opentest4j comparison errors properly to team city output to show diffs in all contexts
* Adds KMP handling for comparison errors
* Moves all the non-shared assertions/helpers into the assertions core module, leaving the namespace unpolluted for anyone not using kotest assertions
 * Unifies types into the Printed typeclass. No custom plugin mechansim exists for Printed typeclasses, so it's safe to make breaking changes
 * Move the long defunct table testing code into its own module and deprecate it all
