+++ kotest-assertions/kotest-assertions-shared/src/commonMain/kotlin/io/kotest/assertions/AssertionErrorBuilder.kt
@@ -1,7 +1,6 @@
 package io.kotest.assertions
 
 import io.kotest.assertions.print.Printed
-import io.kotest.matchers.clueContextAsString
 
 /**
  * Use this object to create exceptions on a target platform.
@@ -58,49 +57,7 @@ data class AssertionErrorBuilder(
    }
 
    fun build(): AssertionError {
-      return createAssertionError(generateMessageString(), cause, expected, actual)
-   }
-
-   private fun generateMessageString(): String {
-      return buildString {
-         append(clueContextAsString())
-         if (message != null) {
-            append(message)
-         }
-         // intellij requires the expected and actual values to be printed in a specific format
-         // as the last part of the message string
-         if (expected != null && actual != null) {
-            append(intellijFormattedComparison(expected, actual))
-         }
-      }
-   }
-
-   /**
-    * Returns a message formatted appropriately for intellij to show a diff.
-    *
-    * This is the format intellij requires in order to recognize the diff:
-    * https://github.com/JetBrains/intellij-community/blob/master/java/java-runtime/src/com/intellij/rt/execution/testFrameworks/AbstractExpectedPatterns.java
-    *
-    * From the above link:
-    * private static final Pattern ASSERT_EQUALS_PATTERN = Pattern.compile("expected:<(.*)> but was:<(.*)>", Pattern.DOTALL | Pattern.CASE_INSENSITIVE);
-    *
-    * Note - if the printed value contains type information and the types are different, then the type
-    * information will be included in the output.
-    */
-   fun intellijFormattedComparison(expected: Expected, actual: Actual): String {
-
-      // only include types if they are different and neither is null
-      val includeTypes = when {
-         expected.value.type == null || actual.value.type == null -> false
-         expected.value.type != actual.value.type -> true
-         else -> false
-      }
-
-      fun format(printed: Printed): String = when (includeTypes) {
-         true if printed.type?.qualifiedName != null -> "${printed.type.qualifiedName}<${printed.value}>"
-         else -> "<${printed.value}>"
-      }
-      return "expected:${format(expected.value)} but was:${format(actual.value)}"
+      return createAssertionError(message, cause, expected, actual)
    }
 }
 
@@ -112,7 +69,7 @@ data class AssertionErrorBuilder(
  * with tools that look for these special exception types to show diffs in the IDE.
  */
 expect fun createAssertionError(
-   message: String,
+   message: String?,
    cause: Throwable?,
    expected: Expected?,
    actual: Actual?
